Added more depth to dialogue
To talk to monsters you now have to be at least 3f close to them and you will have to press [F]. Also to go to a different room you have to be 3f away from them and press [E]. There is also a title screen with different options. Last but not least, there is now a pause menu when you hit "esc", this will also give you different options to choose from. This is more than enough for a toy. I will later expand on dialogue, because that is tricky to work with. I will also add the love meter, save/load to be able to remember what you talked about with the monsters, and a day counter.