Love Meter
There is now a love meter that pops up whenever you start a conversation with a monster. There are also choices at the end of the conversation that change the love meter depending on the answer you choose. The sphere for right now takes you to the Beholders room where there is another monster which you can also interact with. The sphere in that room will take you back to the lounge room. The only thing I do not have is code that allows to branch off to different dialogue depending on your choices and a save/load system.