Setting Up
I have most of it ready for assets to be put in. I just need to fix something with the love meter not showing the progress once you leave the room and come back. It saves the value of the love meter, but for somereason it doesn't show it in the love meter bar.